feat: 根據服務變動調整 getNewTickInfo 的回傳成交量欄位型別並加入單元測試

+++ backend2/src/main/java/com/cmoney/backend2/realtimeaftermarket/service/api/getnewtickinfo/TickInfoSet.kt
@@ -74,13 +74,13 @@ data class TickInfoSet(
      * 本次單量
      */
     @SerializedName("SingleVolume")
-    val singleVolume: Int?,
+    val singleVolume: Long?,
 
     /**
      * 今日總交易量
      */
     @SerializedName("TotalVolume")
-    val totalVolume: Int?,
+    val totalVolume: Long?,
 
     /**
      * 交易時間(UnixTime)
